Set up new code generator, entity module, etc.
