+++ .gitignore
@@ -8,6 +8,7 @@ Cargo.lock
 /extractor/out
 /extractor/classes
 /extractor/run
+/extractor/bin
 rust-mc-bot
 .asset_cache/
 /velocity
