+++ crates/valence_protocol/src/codec.rs
@@ -504,8 +504,8 @@ mod tests {
     use crate::ident::Ident;
     use crate::item::{ItemKind, ItemStack};
     use crate::text::{Text, TextFormat};
-    use crate::tracked_data::PaintingKind;
     use crate::var_long::VarLong;
+    use crate::types::Hand;
     use crate::Decode;
 
     #[cfg(feature = "encryption")]
@@ -520,7 +520,7 @@ mod tests {
         d: f32,
         e: f64,
         f: BlockPos,
-        g: PaintingKind,
+        g: Hand,
         h: Ident<&'a str>,
         i: Option<ItemStack>,
         j: Text,
@@ -540,7 +540,7 @@ mod tests {
                 d: 5.001,
                 e: 1e10,
                 f: BlockPos::new(1, 2, 3),
-                g: PaintingKind::DonkeyKong,
+                g: Hand::Off,
                 h: Ident::new("minecraft:whatever").unwrap(),
                 i: Some(ItemStack::new(ItemKind::WoodenSword, 12, None)),
                 j: "my ".into_text() + "fancy".italic() + " text",
